fastapi/requirements.txt
aiohttp==3.8.6
elasticsearch==8.10.1
fastapi==0.104.0
orjson==3.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
redis==5.0.1
uvicorn==0.23.2
